Basic scenario testing This introduces 3 large-ish end to end tests for Maestro/BAR - Channel management using the darc CLI - GitHub dependency flow (non-batched, without auto merge) - Azure DevOps dependency flow (non-batched, without auto merge)

+++ src/Maestro/Maestro.Web/Api/v2018_07_16/Controllers/ChannelsController.cs
@@ -71,6 +71,12 @@ namespace Maestro.Web.Api.v2018_07_16.Controllers
                 return NotFound();
             }
 
+            // Ensure that there are no subscriptions associated with the channel
+            if (await _context.Subscriptions.AnyAsync(s => s.ChannelId == id))
+            {
+                return Conflict(new ApiError($"Channel {id} has associated subscriptions and cannot be removed."));
+            }
+
             _context.Channels.Remove(channel);
 
             await _context.SaveChangesAsync();
